requirements.txt
Flask==0.12.1
Wand==0.4.4
boto3==1.4.4
PyYAML==3.12
redis==2.10.5
requests==2.14.2
